three fix

+++ philo_two/philo.h
@@ -29,7 +29,7 @@
 uint8_t				g_error;
 uint32_t			g_time;
 uint8_t				g_check_eating;
-uint32_t			g_input_array[5];
+short				g_input_array[5];
 sem_t				*g_semaphore;
 
 typedef struct		s_thread
